refactor(checker): extract base-instance instantiation into its own module (2000-line cap)
constructors.rs exceeded the 2000-line cap after the mint-prevention deferral.
Move instantiate_base_instance_type_with_args[_deferrable] into a dedicated
base_instance_instantiation submodule. No behavior change.
